+++ zinnia/tests/views.py
@@ -342,7 +342,6 @@ class ZinniaViewsTestCase(ViewsBaseCase):
             entry = Entry.objects.create(**params)
             entry.sites.add(self.site)
             entry.categories.add(self.category)
-            entry.authors.add(self.author)
         response = self.client.get('/categories/tests/')
         self.assertEquals(len(response.context['object_list']), PAGINATION)
         response = self.client.get('/categories/tests/?page=2')
@@ -377,6 +376,29 @@ class ZinniaViewsTestCase(ViewsBaseCase):
         self.assertEquals(response.context['author'].username, 'admin')
         restore_template_loaders()
 
+    def test_zinnia_author_detail_paginated(self):
+        """Test case reproducing issue #207 on author
+        detail view paginated"""
+        setup_test_template_loader(
+            {'zinnia/entry_list.html': ''})
+        for i in range(PAGINATION):
+            params = {'title': 'My entry %i' % i,
+                      'content': 'My content %i' % i,
+                      'slug': 'my-entry-%i' % i,
+                      'creation_date': datetime(2010, 1, 1),
+                      'status': PUBLISHED}
+            entry = Entry.objects.create(**params)
+            entry.sites.add(self.site)
+            entry.authors.add(self.author)
+        response = self.client.get('/authors/admin/')
+        self.assertEquals(len(response.context['object_list']), PAGINATION)
+        response = self.client.get('/authors/admin/?page=2')
+        self.assertEquals(len(response.context['object_list']), 2)
+        response = self.client.get('/authors/admin/page/2/')
+        self.assertEquals(len(response.context['object_list']), 2)
+        self.assertEquals(response.context['author'].username, 'admin')
+        restore_template_loaders()
+
     def test_zinnia_tag_list(self):
         setup_test_template_loader(
             {'zinnia/tag_list.html': ''})
@@ -403,6 +425,27 @@ class ZinniaViewsTestCase(ViewsBaseCase):
         self.assertEquals(response.status_code, 404)
         restore_template_loaders()
 
+    def test_zinnia_tag_detail_paginated(self):
+        setup_test_template_loader(
+            {'zinnia/entry_list.html': ''})
+        for i in range(PAGINATION):
+            params = {'title': 'My entry %i' % i,
+                      'content': 'My content %i' % i,
+                      'slug': 'my-entry-%i' % i,
+                      'tags': 'tests',
+                      'creation_date': datetime(2010, 1, 1),
+                      'status': PUBLISHED}
+            entry = Entry.objects.create(**params)
+            entry.sites.add(self.site)
+        response = self.client.get('/tags/tests/')
+        self.assertEquals(len(response.context['object_list']), PAGINATION)
+        response = self.client.get('/tags/tests/?page=2')
+        self.assertEquals(len(response.context['object_list']), 2)
+        response = self.client.get('/tags/tests/page/2/')
+        self.assertEquals(len(response.context['object_list']), 2)
+        self.assertEquals(response.context['tag'].name, 'tests')
+        restore_template_loaders()
+
     def test_zinnia_entry_search(self):
         setup_test_template_loader(
             {'zinnia/entry_search.html': ''})
@@ -416,6 +459,15 @@ class ZinniaViewsTestCase(ViewsBaseCase):
         self.assertEquals(len(response.context['object_list']), 0)
         self.assertEquals(response.context['error'],
                           _('No pattern to search found'))
+        restore_template_loaders()
+
+    def test_zinnia_entry_random(self):
+        setup_test_template_loader(
+            {'zinnia/entry_detail.html': ''})
+        response = self.client.get('/random/', follow=True)
+        self.assertTrue(response.redirect_chain[0][0].startswith(
+            'http://testserver/2010/'))
+        self.assertEquals(response.redirect_chain[0][1], 302)
         restore_template_loaders()
 
     def test_zinnia_sitemap(self):